requirements.txt
Django>=5.0,<5.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
python-dotenv==1.0.1
PyMySQL==1.1.0
gunicorn==21.2.0
asgiref==3.11.0
sqlparse==0.5.5
tzdata==2025.3
python-decouple==3.8

cryptography>=41.0.0
